// src/type/type.ts
export interface Equipment {
    uid: string;
    active: boolean;
    brand: string;
    createdAt: Date;
    createdBy: string;
    department: string;
    hardDisk: string;
    ip: string;
    model: string;
    observations: string;
    officer: string;
    photosURL: string[];
    ram: string;
    serie: string;
    type: string;
    ubication: string;
    userAssigned: string;
}